allow to stop PTU via speed command
To allow stopping the ptu at the current position
a joint command with speed set to zero can now be used.

In this case the hardware ignores that the speed is set to zero
and we have to call tiltStop and panStop ourself

Furthermore, if the speed is not set the driver restores the
speed defined by the property. Otherwise it uses the one from the
command.

Speed only commands are not supported.